requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
boto3==1.42.64
botocore==1.42.64
frozenlist==1.8.0
idna==3.11
jmespath==1.1.0
multidict==6.7.1
propcache==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
s3transfer==0.16.0
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0
